fix new component
