+++ src/app/posts/graphql/queries.ts
@@ -1,9 +1,4 @@
 import gql from 'graphql-tag';
-export interface getPostDataByIdInterface {
-    id:string,
-    title:string | null,
-    content:string | null
-}
 
 export const GetPostsQuery = gql`
   query Posts {
