+++ src/app/posts/new-post/new-post.component.html
@@ -2,11 +2,11 @@
     <h5>Add new post:</h5>
     <form [formGroup]="form" (ngSubmit)="save()">
         <md-input-container>
-            <input mdInput formControlName="title" placeholder="Title" [(ngModel)]="title">
+            <input mdInput formControlName="title" placeholder="Title">
             <md-error>This field is required</md-error>
         </md-input-container><br>
         <md-input-container>
-            <textarea mdInput md-autosize minRows="6" placeholder="Your post content..." formControlName="content" [(ngModel)]="content"></textarea>
+            <textarea mdInput md-autosize minRows="6" placeholder="Your post content..." formControlName="content"></textarea>
         </md-input-container>
         <button md-raised-button type="submit">Save</button>
     </form>
