+++ src/app/posts/posts.module.ts
@@ -6,7 +6,7 @@ import { HttpModule }  from '@angular/http';
 import { MaterialModule } from '@angular/material';
 
 import { routes } from './posts.routes';
-import { PostsComponent } from './posts.component';
+import { PostListComponent } from './post-list/post-list.component';
 import { postDetailComponent } from './post-detail/post-detail.component';
 import { NewPostComponent } from './new-post/new-post.component';
 import {PostsFilterPipe} from './posts-filter/posts-filter.pipe'
@@ -16,7 +16,7 @@ import {PostsFilterPipe} from './posts-filter/posts-filter.pipe'
     /**
      * Components / Directives/ Pipes
      */
-    PostsComponent,
+    PostListComponent,
     NewPostComponent,
     PostsFilterPipe,
     postDetailComponent,
